+++ packages/create-gasket-app/lib/scaffold/actions/load-preset.js
@@ -1,82 +1,60 @@
-const path = require('path');
-const action = require('../action-wrapper');
-const PackageFetcher = require('../fetcher');
-const { presetIdentifier, Loader } = require('@gasket/resolve');
-
-const loader = new Loader();
-
-/**
- * Fetches the preset packages and loads PresetInfos
- *
- * @param {[String]} rawPresets - Presets names
- * @param {String} cwd - Root path
- * @param {String} from from
- * @returns {PresetInfo[]} loaded presetInfos
- * @private
- */
-async function remotePresets(rawPresets, cwd, from) {
-  if (!rawPresets) { return []; }
-
-  const allRemotePresets = await Promise.all(rawPresets.map(async rawName => {
-    const packageName = presetIdentifier(rawName).full;
-    const fetcher = new PackageFetcher({ cwd, packageName });
-    const pkgPath = await fetcher.clone();
-
-    const presetInfo = loader.loadPreset(pkgPath, { from: from, rawName }, { shallow: true });
-    const { name: presetName, dependencies } = presetInfo.package;
-    if (!dependencies) {
-      return presetInfo;
-    }
-
-    const presetDeps = Object.keys(dependencies).filter(k => presetIdentifier.isValidFullName(k));
-    const rawPresetsDeps = presetDeps.map(presetDep => `${presetDep}@${dependencies[presetDep]}`);
-
-    const presetInfoDeps = await remotePresets(rawPresetsDeps, cwd, presetName);
-    return {
-      ...presetInfo,
-      presets: presetInfoDeps
-    };
-  }));
-
-  return allRemotePresets;
-}
+import path from 'path';
+import os from 'os';
+import action from '../action-wrapper.js';
+import { default as gasketUtils } from '@gasket/utils';
+import { mkdtemp } from 'fs/promises';
+import { createRequire } from 'module';
+const require = createRequire(import.meta.url);
+const hasVersion = /@(\^[0-9]\.|[0-9]).+/;
 
 /**
- * Loads PresetInfo from local preset package
- *
+ * loadPresets - Load presets to temp directory
  * @param {CreateContext} context - Create context
- * @returns {PresetInfo[]} loaded presetInfos
- * @private
  */
-function localPreset(context) {
-  const { cwd, localPresets = [] } = context;
-  if (!localPresets) { return []; }
+async function loadPresets({ context }) {
+  const tmpDir = await mkdtemp(path.join(os.tmpdir(), `gasket-create-${context.appName}`));
+  context.tmpDir = tmpDir;
 
-  const presetInfos = [];
-  localPresets.forEach(localPresetPath => {
-    const pkgPath = path.resolve(cwd, localPresetPath);
-    const presetInfo = loader.loadPreset(pkgPath, { from: 'cli' }, { shallow: false });
-    presetInfo.rawName = `${presetInfo.package.name}@file:${pkgPath}`;
-    presetInfos.push(presetInfo);
+  const modPath = path.join(tmpDir, 'node_modules');
+  const pkgManager = new gasketUtils.PackageManager({
+    packageManager: context.packageManager,
+    dest: tmpDir
+  });
+  const pkgVerb = pkgManager.isYarn ? 'add' : 'install';
+
+  const remotePresets = context.rawPresets.map(async preset => {
+    const parts = hasVersion.test(preset) && preset.split('@').filter(Boolean);
+    const name = parts ? `@${parts[0]}` : preset;
+    const version = parts ? `@${parts[1]}` : '@latest';
+
+    try {
+      // Install the preset
+      await pkgManager.exec(pkgVerb, [`${name}${version}`]);
+      // Get the package file
+      const pkgFile = require(path.join(modPath, name, 'package.json'));
+      // Import the preset via the package file attrs, name and main
+      // We can't specify the cwd for the import, so we need to use the full path
+      // expects type:module & "main": "lib/fullpath.js"
+      const mod = await import(`${modPath}/${name}/${pkgFile.main}`);
+      return mod.default || mod;
+    } catch (err) {
+      throw new Error(`Failed to install preset ${name}${version}`);
+    }
   });
 
-  return presetInfos;
+  const localPresets = context.localPresets.map(async localPresetPath => {
+
+    try {
+      await pkgManager.exec(pkgVerb, [localPresetPath]);
+      const pkgFile = require(path.join(localPresetPath, 'package.json'));
+      const mod = await import(`${modPath}/${pkgFile.name}/${pkgFile.main}`);
+      return mod.default || mod;
+    } catch (err) {
+      throw new Error(`Failed to install local preset ${localPresetPath}`);
+    }
+  });
+
+  context.presets = await Promise.all([...remotePresets, ...localPresets]);
 }
 
-/**
- * Downloads the target preset package and adds the package.json
- * contents and plugins dependencies to context.
- *
- * @param {CreateContext} context - Create context
- * @returns {Promise} promise
- */
-async function loadPreset(context) {
-  const { rawPresets = [], cwd } = context;
-  let presetInfos = await remotePresets(rawPresets, cwd, 'cli');
-  presetInfos = presetInfos.concat(localPreset(context));
-
-  const presets = presetInfos.map(p => presetIdentifier(p.rawName).shortName);
-  Object.assign(context, { presets, presetInfos });
-}
-
-module.exports = action('Load presets', loadPreset);
+export default action('Load presets', loadPresets);