+++ packages/create-gasket-app/lib/scaffold/actions/prompt-hooks.js
@@ -1,47 +1,5 @@
-const inquirer = require('inquirer');
-const action = require('../action-wrapper');
-const { addPluginsToContext, addPluginsToPkg, getPluginsWithVersions } = require('../utils');
-const createEngine = require('../create-engine');
-
-/**
- * Create the `addPlugins` function with context
- *
- * @param {CreateContext} context - Create context
- * @returns {addPlugins} addPlugins
- */
-const createAddPlugins = context => {
-  /**
-   * Allows plugins to add plugins based on user input.
-   *
-   * @typedef {Function} addPlugins
-   *
-   * @param {...PluginDesc} pluginsToAdd - Additional plugins to add
-   * @returns {Promise} promise
-   */
-  return async function addPlugins(...pluginsToAdd) {
-    const { pkg, pkgLinks = [], pkgManager } = context;
-
-    addPluginsToContext(pluginsToAdd, context);
-    const pluginIds = await getPluginsWithVersions(pluginsToAdd, pkgManager);
-    addPluginsToPkg(pluginIds, pkg);
-
-    //
-    // Install new plugins if not already linked
-    //
-    const toInstall = pluginIds.filter(p => !pkgLinks.includes(p.fullName));
-    if (toInstall.length) {
-      await pkgManager.install(toInstall.map(p => p.full));
-      //
-      // relink because npm blows them away on installs...
-      //
-      if (pkgLinks.length) {
-        await pkgManager.link(pkgLinks);
-      }
-    }
-
-    await execPluginPrompts(context, pluginIds.map(p => p.shortName));
-  };
-};
+import inquirer from 'inquirer';
+import action from '../action-wrapper.js';
 
 /**
  * Initializes engine with provide preset and plugins
@@ -53,18 +11,12 @@ const createAddPlugins = context => {
  * @returns {Promise} promise
  * @private
  */
-async function execPluginPrompts(context, plugins = [], presets = []) {
-  const { dest } = context;
-
-  const gasket = await createEngine({ dest, presets, plugins });
-
+async function execPluginPrompts(gasket, context) {
   //
   // @see: https://github.com/SBoudrias/Inquirer.js/#inquirercreatepromptmodule---prompt-function
   //
   const prompt = context.prompts ? inquirer.createPromptModule() : () => ({});
-  const addPlugins = createAddPlugins(context);
-
-  const nextContext = await gasket.execWaterfall('prompt', context, { prompt, addPlugins });
+  const nextContext = await gasket.execWaterfall('prompt', context, { prompt });
   //
   // Ensure the original context is transformed in case a prompt returns a new object
   //
@@ -73,18 +25,19 @@ async function execPluginPrompts(context, plugins = [], presets = []) {
 
 /**
  * Executes the `prompt` hook for all registered plugins.
- * Adds `prompt` and `addPlugins` util functions for prompting features.
+ * Adds `prompt` util function for prompting features.
  *
+ * @param {GasketEngine} gasket - Gasket API
  * @param {CreateContext} context - Create context
  * @returns {Promise} promise
  */
-async function promptHooks(context) {
+async function promptHooks({ gasket, context }) {
   //
   // Because `execPluginPrompts` is recursively, we need to start it
   // with the processPlugins and presets from our initial context
   //
-  const { presets, plugins } = context;
-  await execPluginPrompts(context, plugins, presets);
+  const { plugins } = context;
+  await execPluginPrompts(gasket, context, plugins);
 }
 
-module.exports = action('Plugin prompts', promptHooks, { startSpinner: false });
+export default action('Plugin prompts', promptHooks, { startSpinner: false });