+++ packages/gasket-plugin-express/generator/routes/index.js
@@ -1,23 +1,25 @@
-module.exports = (app) => {
-  {{#if hasSwaggerPlugin}}
-  /**
-  * @swagger
-  *
-  * /default:
-  *   get:
-  *     summary: "Get default route"
-  *     produces:
-  *       - "application/json"
-  *     responses:
-  *       "200":
-  *         description: "Returns welcome message."
-  *         content:
-  *           application/json
-  */
-  {{/if}}
-  app.get('/default', async (req, res) => {
-    res.status(200).json({
-      message: 'Welcome to your default route...'
+export const routes = [
+  (app) => {
+    {{#if hasSwaggerPlugin}}
+    /**
+    * @swagger
+    *
+    * /default:
+    *   get:
+    *     summary: "Get default route"
+    *     produces:
+    *       - "application/json"
+    *     responses:
+    *       "200":
+    *         description: "Returns welcome message."
+    *         content:
+    *           application/json
+    */
+    {{/if}}
+    app.get('/default', async (req, res) => {
+      res.status(200).json({
+        message: 'Welcome to your default route...'
+      });
     });
-  });
-};
+  }
+];